+++ lib/presentation/pages/home_page/problem_path/header/progress_header.dart
@@ -1,0 +1,50 @@
+import 'package:algopath_app/config/res/app_dimens.dart';
+import 'package:flutter/material.dart';
+
+class ProgressHeader extends StatelessWidget {
+  const ProgressHeader({
+    super.key,
+    required this.completed,
+    required this.total,
+  });
+
+  final int completed;
+  final int total;
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: const EdgeInsets.all(AppDimens.defaultPadding),
+      child: Column(
+        crossAxisAlignment: CrossAxisAlignment.start,
+        children: [
+          Text(
+            'Grind 75 questions',
+            style: Theme.of(context).textTheme.headlineLarge?.copyWith(
+                  fontWeight: FontWeight.bold,
+                ),
+          ),
+          const SizedBox(height: 8),
+          Text(
+            'Customize LeetCode study plans according to your needs. You are recommended to work on the questions in order.',
+            style: Theme.of(context).textTheme.bodyLarge?.copyWith(
+                  color: Colors.grey[600],
+                ),
+          ),
+          const SizedBox(height: 24),
+          LinearProgressIndicator(
+            value: completed / total,
+            backgroundColor: Colors.grey[200],
+            borderRadius: BorderRadius.circular(4),
+          ),
+          const SizedBox(height: 8),
+          Text(
+            'COMPLETED $completed / $total',
+            style: Theme.of(context).textTheme.bodyMedium?.copyWith(
+                  fontWeight: FontWeight.w500,
+                ),
+          ),
+        ],
+      ),
+    );
+  }
+}
