+++ lib/presentation/pages/home_page/problem_path/panel/right_info_panel.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+class RightInfoPanel extends StatelessWidget {
+  const RightInfoPanel({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return const SizedBox();
+  }
+}
